requirements.txt
django
mysqlclient
djangorestframework
requests